+++ api/src/utils.js
@@ -18,6 +18,16 @@ function validatePassword(password) {
   return schema.validate(password);
 }
 
+function generatePassword() {
+  let length = 6;
+  let charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+  let retVal = "";
+  for (let i = 0, n = charset.length; i < length; ++i) {
+    retVal += charset.charAt(Math.floor(Math.random() * n));
+  }
+  return retVal;
+}
+
 async function comparePassword(password, expected) {
   return bcrypt.compare(password, expected);
 }
@@ -30,4 +40,5 @@ module.exports = {
   validatePassword,
   comparePassword,
   hashPassword,
+  generatePassword,
 };