+++ dashboard/src/components/drawer.js
@@ -23,11 +23,12 @@ const Drawer = () => {
   const [currentTeam, setCurrentTeam] = useRecoilState(currentTeamState);
   const API = useApi();
 
+  const onboardingForEncryption = !organisation.encryptionEnabled;
   const onboardingForTeams = !teams.length;
 
   return (
     <>
-      <Sidebar className="noprint" onboardingForTeams={onboardingForTeams}>
+      <Sidebar className="noprint" isOnboarding={onboardingForEncryption || onboardingForTeams}>
         <Nav>
           {!['superadmin'].includes(user.role) && (
             <>
@@ -52,14 +53,14 @@ const Drawer = () => {
           )}
           <hr />
           {['admin'].includes(user.role) && (
-            <li>
+            <li id="show-on-onboarding">
               <NavLink to={`/organisation/${organisation._id}`} activeClassName="active">
                 Organisation
               </NavLink>
             </li>
           )}
           {['admin'].includes(user.role) && (
-            <li id="teams">
+            <li id="show-on-onboarding">
               <NavLink to="/team" activeClassName="active">
                 Équipes
               </NavLink>
@@ -214,9 +215,9 @@ const Sidebar = styled.div`
   border-right: 1px solid rgba(0, 0, 0, 0.1);
 
   ${(p) =>
-    p.onboardingForTeams &&
+    p.isOnboarding &&
     `
-    li:not(#teams) {
+    li:not(#show-on-onboarding) {
       opacity: 0.2;
       pointer-events: none;
     }