+++ dashboard/src/scenes/team/list.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { Col, Container, FormGroup, Input, Modal, ModalBody, ModalHeader, Row } from 'reactstrap';
+import { Col, Container, FormGroup, Input, Label, Modal, ModalBody, ModalHeader, Row } from 'reactstrap';
 import { useHistory } from 'react-router-dom';
 import { Formik } from 'formik';
 import { toastr } from 'react-redux-toastr';
@@ -15,6 +15,7 @@ import NightSessionModale from '../../components/NightSessionModale';
 import { currentTeamState, organisationState, teamsState, userState } from '../../recoil/auth';
 import useApi from '../../services/api';
 import { AppSentry } from '../../services/sentry';
+import OnboardingEndModal from '../../components/OnboardingEndModal';
 
 const List = () => {
   const teams = useRecoilValue(teamsState);
@@ -56,14 +57,17 @@ const Create = () => {
   const setCurrentTeam = useSetRecoilState(currentTeamState);
   const [open, setOpen] = useState(!teams.length);
   const API = useApi();
+  const [onboardingEndModalOpen, setOnboardingEndModalOpen] = useState(false);
 
   const onboardingForTeams = !teams.length;
 
   return (
     <CreateWrapper>
       <ButtonCustom color="primary" onClick={() => setOpen(true)} title="Créer une nouvelle équipe" padding="12px 24px" />
-      <Modal isOpen={open} toggle={() => setOpen(false)} size="lg">
-        <ModalHeader toggle={() => setOpen(false)}>{onboardingForTeams ? 'Bienvenue dans Mano !' : 'Créer une nouvelle équipe'}</ModalHeader>
+      <Modal isOpen={open} toggle={() => setOpen(false)} size="lg" backdrop={onboardingForTeams ? 'static' : true}>
+        <ModalHeader close={onboardingForTeams ? <></> : null} toggle={() => setOpen(false)}>
+          {onboardingForTeams ? 'Dernière étape !' : 'Créer une nouvelle équipe'}
+        </ModalHeader>
         <ModalBody>
           <span>Veuillez créer une première équipe avant de commencer à utiliser la plateforme</span>
           <br />
@@ -71,7 +75,15 @@ const Create = () => {
           <Formik
             initialValues={{ name: '' }}
             onSubmit={async (values, actions) => {
-              const newTeamRes = await API.post({ path: '/team', body: { name: values.name, organisation: organisation._id } });
+              if (!values.name) {
+                toastr.error('Vous devez choisir un nom');
+                actions.setSubmitting(false);
+                return;
+              }
+              const newTeamRes = await API.post({
+                path: '/team',
+                body: { name: values.name, organisation: organisation._id, nightSession: values.nightSession === 'true' },
+              });
               if (!newTeamRes.ok) return actions.setSubmitting(false);
               if (onboardingForTeams) {
                 const userPutRes = await API.put({ path: `/user/${user._id}`, body: { team: [newTeamRes.data._id] } });
@@ -81,6 +93,7 @@ const Create = () => {
                 AppSentry.setUser(meResponse.user);
                 setCurrentTeam(meResponse.user.teams[0]);
                 toastr.success('Création réussie !', `Vous êtes dans l'équipe ${newTeamRes.data.name}`);
+                setOnboardingEndModalOpen(true);
               } else {
                 toastr.success('Création réussie !');
               }
@@ -95,8 +108,39 @@ const Create = () => {
                   <Row>
                     <Col md={6}>
                       <FormGroup>
-                        <div>Nom</div>
+                        <Label>Nom</Label>
                         <Input name="name" value={values.name} onChange={handleChange} />
+                      </FormGroup>
+                    </Col>
+                    <Col md={6}>
+                      <FormGroup>
+                        <Label>L'équipe travaille-t-elle de nuit ?</Label>
+                        <div style={{ display: 'flex', flexDirection: 'column', marginLeft: 20, width: '80%' }}>
+                          <FormGroup style={{ marginBottom: 0 }}>
+                            <input
+                              style={{ marginRight: 10 }}
+                              type="radio"
+                              id="nightSessionYes"
+                              name="nightSession"
+                              value="true"
+                              checked={values.nightSession === 'true'}
+                              onChange={handleChange}
+                            />
+                            <label htmlFor="nightSessionYes">Oui</label>
+                          </FormGroup>
+                          <FormGroup style={{ marginBottom: 0 }}>
+                            <input
+                              style={{ marginRight: 10 }}
+                              type="radio"
+                              id="nightSessionNo"
+                              name="nightSession"
+                              value="false"
+                              checked={values.nightSession === 'false'}
+                              onChange={handleChange}
+                            />
+                            <label htmlFor="nightSessionNo">Non</label>
+                          </FormGroup>
+                        </div>
                       </FormGroup>
                     </Col>
                   </Row>
@@ -112,6 +156,7 @@ const Create = () => {
           </Formik>
         </ModalBody>
       </Modal>
+      <OnboardingEndModal open={onboardingEndModalOpen} setOpen={setOnboardingEndModalOpen} />
     </CreateWrapper>
   );
 };